♻️ Add feature flag for the DEX feature
