+++ lib/domain/repositories/features_flags.dart
@@ -2,4 +2,5 @@
 class FeatureFlags {
   static const messagingActive = false;
   static const forceLogout = false;
+  static const dexActive = true;
 }
