+++ lib/ui/views/sheets/unavailable_feature_warning.dart
@@ -1,0 +1,86 @@
+/// SPDX-License-Identifier: AGPL-3.0-or-later
+
+import 'package:aewallet/application/settings/settings.dart';
+import 'package:aewallet/ui/themes/styles.dart';
+import 'package:aewallet/ui/util/dimens.dart';
+import 'package:aewallet/ui/views/main/components/sheet_appbar.dart';
+import 'package:aewallet/ui/widgets/components/app_button_tiny.dart';
+import 'package:aewallet/ui/widgets/components/sheet_skeleton.dart';
+import 'package:aewallet/ui/widgets/components/sheet_skeleton_interface.dart';
+import 'package:aewallet/util/get_it_instance.dart';
+import 'package:aewallet/util/haptic_util.dart';
+import 'package:flutter/material.dart';
+import 'package:flutter_gen/gen_l10n/localizations.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_vibrate/flutter_vibrate.dart';
+import 'package:go_router/go_router.dart';
+import 'package:material_symbols_icons/symbols.dart';
+
+class UnavailableFeatureWarning extends ConsumerWidget
+    implements SheetSkeletonInterface {
+  const UnavailableFeatureWarning({super.key});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    return SheetSkeleton(
+      appBar: getAppBar(context, ref),
+      floatingActionButton: getFloatingActionButton(context, ref),
+      sheetContent: getSheetContent(context, ref),
+    );
+  }
+
+  @override
+  Widget getFloatingActionButton(BuildContext context, WidgetRef ref) {
+    final preferences = ref.watch(SettingsProviders.settings);
+    return Row(
+      children: <Widget>[
+        AppButtonTiny(
+          AppButtonTinyType.primary,
+          AppLocalizations.of(context)!.understandButton,
+          Dimens.buttonBottomDimens,
+          key: const Key('Understand'),
+          onPressed: () async {
+            sl.get<HapticUtil>().feedback(
+                  FeedbackType.light,
+                  preferences.activeVibrations,
+                );
+            context.pop();
+          },
+        ),
+      ],
+    );
+  }
+
+  @override
+  PreferredSizeWidget getAppBar(BuildContext context, WidgetRef ref) {
+    return SheetAppBar(
+      title: AppLocalizations.of(context)!.information,
+    );
+  }
+
+  @override
+  Widget getSheetContent(BuildContext context, WidgetRef ref) {
+    final localizations = AppLocalizations.of(context)!;
+    return Column(
+      children: <Widget>[
+        Row(
+          children: [
+            const Icon(Symbols.warning, color: Colors.red),
+            const SizedBox(width: 10),
+            Text(
+              localizations.webChannelIncompatibilityWarning,
+              style: ArchethicThemeStyles.textStyleSize14W600Primary,
+            ),
+          ],
+        ),
+        const SizedBox(
+          height: 30,
+        ),
+        Text(
+          localizations.webChannelIncompatibilityWarningDesc,
+          style: ArchethicThemeStyles.textStyleSize12W100Primary,
+        ),
+      ],
+    );
+  }
+}
